requirements.txt
# Core ML/DL dependencies
torch>=2.2.2
torchvision>=0.17.2
torchaudio>=2.2.2
numpy>=1.26.0
scipy>=1.4.1

# Data manipulation and analysis
pandas>=1.1.4
PyYAML>=6.0.2

# Visualization
matplotlib>=3.9.2
seaborn>=0.11.0
opencv-python>=4.1.1
pillow>=10.4.0

# Development tools
ipykernel>=6.29.5
ipython>=8.27.0
jupyter_client>=8.6.2
jupyter_core>=5.7.2

# Utility packages
tqdm>=4.66.6
requests>=2.32.3
pathlib>=1.0.1
filelock>=3.16.0
psutil>=6.0.0

# Optional dependencies that might be needed by other packages
packaging>=24.1
typing_extensions>=4.12.2
urllib3>=2.2.3
sympy>=1.13.2
networkx>=3.3